Refactor HttpContext import and location
- Move HttpContext from base to data directory
- Update import statements in Middleware and related files
- Maintain consistent import paths for HttpContext
